+++ src/lib/get-initialized-liff.ts
@@ -1,15 +1,14 @@
-import { Liff, LiffPlugin } from '@line/liff';
+import { Liff, LiffPlugin } from "@line/liff";
 
-// eslint-disable-next-line filenames-simple/typescript-module-declaration
 declare global {
   interface Window {
     liff?: Liff;
   }
 }
 
-type LiffConfig = Parameters<Liff['init']>[0];
+type LiffConfig = Parameters<Liff["init"]>[0];
 
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
+// biome-ignore lint/suspicious/noExplicitAny:
 type Plugin<PluginOption = any> = LiffPlugin<any, void> | [LiffPlugin<any, PluginOption>, PluginOption];
 
 interface GetInitializedLiffProps extends LiffConfig {
@@ -20,9 +19,8 @@ interface GetInitializedLiffProps extends LiffConfig {
 type GetInitializedLiff = (props: GetInitializedLiffProps) => Promise<Liff>;
 
 const getLiff = async (): Promise<Liff> => {
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
   // @ts-ignore: This is an issue of @line/liff
-  return window.liff ?? ((await import('@line/liff')).default as Liff);
+  return window.liff ?? ((await import("@line/liff")).default as Liff);
 };
 
 const registerLiffPlugin = (liff: Liff, plugin: Plugin) => {
@@ -32,7 +30,9 @@ const registerLiffPlugin = (liff: Liff, plugin: Plugin) => {
 const getInitializedLiff: GetInitializedLiff = async ({ plugins = [], callback = () => {}, ...liffConfig }) => {
   const liff = await getLiff();
 
-  plugins.forEach(plugin => registerLiffPlugin(liff, plugin));
+  for (const plugin of plugins) {
+    registerLiffPlugin(liff, plugin);
+  }
   await liff.init(liffConfig);
   await callback(liff);